strip out file watching
BREAKING CHANGE
File watching technique should not be mandated by the bundler.
File watching will exist in cara-cdn instead, which you are not required to use.
The `watch` Project option is removed.
The `watch` Package method is removed.
New `hold` and `drop` Package methods are added. The bundler emits an event when a package becomes used or unused. The `hold` method should be called once per bundle using the package. When the bundle stops using a package, the `drop` method must be called.